+++ core/server/api_container/server/api_container_service.go
@@ -185,38 +185,36 @@ func (apicService ApiContainerService) ExecuteStartosisScript(ctx context.Contex
 
 	// TODO(gb): add metric tracking maybe?
 
-	interpretationOutput, potentialInterpretationError, generatedInstructionsList, err :=
+	interpretationOutput, potentialInterpretationError, generatedInstructionsList :=
 		apicService.startosisInterpreter.Interpret(ctx, serializedStartosisScript)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "Unable to interpret Startosis script for an unexpected reason: \n%v",
-			serializedStartosisScript)
-	}
 	if potentialInterpretationError != nil {
-		return &kurtosis_core_rpc_api_bindings.ExecuteStartosisScriptResponse{
-			SerializedScriptOutput: interpretationOutput.Get(),
-			InterpretationError:    potentialInterpretationError.Get(),
-		}, nil
+		return binding_constructors.NewExecuteStartosisScriptResponse(
+			string(interpretationOutput),
+			potentialInterpretationError.Error(),
+			"",
+			"",
+		), nil
 	}
-
 	logrus.Debugf("Successfully interpreted Startosis script into a series of Kurtosis instructions: \n%v",
 		generatedInstructionsList)
 
-	executionError, err := apicService.startosisExecutor.Execute(ctx, generatedInstructionsList)
+	err := apicService.startosisExecutor.Execute(ctx, generatedInstructionsList)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "Unexpected error encountered running the Kurtosis instructions: \n%v",
-			generatedInstructionsList)
-	}
-	if executionError != nil {
-		return &kurtosis_core_rpc_api_bindings.ExecuteStartosisScriptResponse{
-			SerializedScriptOutput: interpretationOutput.Get(),
-			ExecutionError:         executionError.Error,
-		}, nil
+		return binding_constructors.NewExecuteStartosisScriptResponse(
+			string(interpretationOutput),
+			"",
+			"",
+			err.Error(),
+		), nil
 	}
 	logrus.Debugf("Successfully executed the list of Kurtosis instructions")
 
-	return &kurtosis_core_rpc_api_bindings.ExecuteStartosisScriptResponse{
-		SerializedScriptOutput: interpretationOutput.Get(),
-	}, nil
+	return binding_constructors.NewExecuteStartosisScriptResponse(
+		string(interpretationOutput),
+		"",
+		"",
+		"",
+	), nil
 }
 
 func (apicService ApiContainerService) StartServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServicesArgs) (*kurtosis_core_rpc_api_bindings.StartServicesResponse, error) {