+++ core/server/api_container/server/startosis_engine/startosis_executor.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_instruction"
 	"github.com/kurtosis-tech/stacktrace"
+	"sync"
 )
 
 type StartosisExecutor struct {
+	mutex *sync.Mutex
 }
 
 type ExecutionError struct {
@@ -14,14 +16,22 @@ type ExecutionError struct {
 }
 
 func NewStartosisExecutor() *StartosisExecutor {
-	// TODO(gb): Implement the bindings to send instructions straight to the backend
-	return &StartosisExecutor{}
+	return &StartosisExecutor{
+		mutex: &sync.Mutex{},
+	}
 }
 
 // Execute executes the list of Kurtosis instructions against the Kurtosis backend
 // It returns a potential execution error if something went wrong.
-// It returns a error if something unexpected happens outside the execution of the script
-func (executor *StartosisExecutor) Execute(ctx context.Context, instructions []kurtosis_instruction.KurtosisInstruction) (*ExecutionError, error) {
-	// TODO(gb): implement
-	return nil, stacktrace.NewError("not implemented")
+// It returns an error if something unexpected happens outside the execution of the script
+func (executor *StartosisExecutor) Execute(ctx context.Context, instructions []kurtosis_instruction.KurtosisInstruction) error {
+	executor.mutex.Lock()
+	defer executor.mutex.Unlock()
+	for index, instruction := range instructions {
+		err := instruction.Execute(ctx)
+		if err != nil {
+			return stacktrace.Propagate(err, "An error occurred executing instruction number '%v'", index)
+		}
+	}
+	return nil
 }